oscar_repo Primer intento de conectar con redux a sequelize y popular BD local, Oscar
