+++ client/src/redux/actions/index.js
@@ -1,0 +1,2 @@
+export const ADD_VIDEOGAME = "ADD_VIDEOGAME";
+export const POST_ARTIST = "POST_ARTIST";
