+++ client/src/redux/actions/postartist.js
@@ -2,7 +2,7 @@ import axios from "axios";
 
 export default function postartist(payload) {
   return async function (dispatch) {
-    var result = await axios.post("http://localhost:3001/artists", payload);
+    var result = await axios.post("http://localhost:3001/artist", payload);
     return result;
   };
 }
