+++ client/src/redux/reducer/index.js
@@ -1,4 +1,4 @@
-import { GET_ARTISTS } from "../actions";
+import { POST_ARTIST } from "../actions";
 
 const initialState = {
   artists: [],
@@ -6,7 +6,7 @@ const initialState = {
 
 export default function rootReducer(state = initialState, action) {
   switch (action.type) {
-    case GET_ARTISTS:
+    case POST_ARTIST:
       if (action.payload) {
         return {
           ...state,
